rewrite tests for DispatcherMiddleman
this will allow the tests to have access to private variables
